Fix whitespace collapse in Chrome 25
Work around a recent change to Chrome that results in the previous
whitespace collapsing technique failing. Chrome doesn't need the
`word-spacing` hack but it has started to respect the property, and this
is causing button group items to overlap.

However, Opera for Windows does need the `word-spacing` hack. The fix is
to use an Opera prefixed pseudo-class in a new selector block for the
`word-spacing` declaration. All other supported browsers will ignore the
rule. Only the Android 4.2 browser still seems to have problems fully
collapsing the whitespace.